Move screenshot observable handler
