+++ x-pack/plugins/screenshotting/server/screenshots/observable.test.ts
@@ -5,494 +5,98 @@
  * 2.0.
  */
 
-import { of, throwError, NEVER } from 'rxjs';
+import { interval, throwError, of } from 'rxjs';
+import { map } from 'rxjs/operators';
 import type { Logger } from 'src/core/server';
+import type { ConditionalHeaders } from '../browsers';
 import { createMockBrowserDriver } from '../browsers/mock';
-import type { HeadlessChromiumDriverFactory } from '../browsers';
-import * as Layouts from '../layouts/create_layout';
 import { createMockLayout } from '../layouts/mock';
-import { CONTEXT_ELEMENTATTRIBUTES } from './constants';
-import { getScreenshots, ScreenshotOptions } from './observable';
+import { ScreenshotObservableHandler, ScreenshotObservableOptions } from './observable';
 
-/*
- * Tests
- */
-describe('Screenshot Observable Pipeline', () => {
-  let driver: ReturnType<typeof createMockBrowserDriver>;
-  let driverFactory: jest.Mocked<HeadlessChromiumDriverFactory>;
+describe('ScreenshotObservableHandler', () => {
+  let browser: ReturnType<typeof createMockBrowserDriver>;
   let layout: ReturnType<typeof createMockLayout>;
   let logger: jest.Mocked<Logger>;
-  let options: ScreenshotOptions;
+  let options: ScreenshotObservableOptions;
 
   beforeEach(async () => {
-    driver = createMockBrowserDriver();
-    driverFactory = {
-      createPage: jest.fn(() => of({ driver, exit$: NEVER })),
-    } as unknown as typeof driverFactory;
+    browser = createMockBrowserDriver();
     layout = createMockLayout();
-    logger = {
-      debug: jest.fn(),
-      error: jest.fn(),
-      info: jest.fn(),
-    } as unknown as jest.Mocked<Logger>;
+    logger = { error: jest.fn() } as unknown as jest.Mocked<Logger>;
     options = {
-      browserTimezone: 'UTC',
-      conditionalHeaders: {},
-      layout: {},
-      timeouts: {
-        loadDelay: 2000,
-        openUrl: 120000,
-        waitForElements: 20000,
-        renderComplete: 10000,
+      conditionalHeaders: {
+        headers: { testHeader: 'testHeadValue' },
+        conditions: {} as unknown as ConditionalHeaders['conditions'],
       },
-      urls: ['/welcome/home/start/index.htm'],
-    } as unknown as typeof options;
+      timeouts: {
+        loadDelay: 5000,
+        openUrl: 30000,
+        waitForElements: 30000,
+        renderComplete: 30000,
+      },
+      urls: [],
+    };
 
-    jest.spyOn(Layouts, 'createLayout').mockReturnValue(layout);
-
-    driver.isPageOpen.mockReturnValue(true);
+    browser.isPageOpen.mockReturnValue(true);
   });
 
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
+  describe('waitUntil', () => {
+    it('catches TimeoutError and references the timeout config in a custom message', async () => {
+      const screenshots = new ScreenshotObservableHandler(browser, logger, layout, options);
+      const test$ = interval(1000).pipe(screenshots.waitUntil(200, 'Test Config'));
 
-  it('pipelines a single url into screenshot and timeRange', async () => {
-    const result = await getScreenshots(driverFactory, logger, options).toPromise();
-
-    expect(result).toMatchInlineSnapshot(`
-      Object {
-        "layout": PreserveLayout {
-          "groupCount": 1,
-          "hasFooter": true,
-          "hasHeader": true,
-          "height": 100,
-          "id": "preserve_layout",
-          "scaledHeight": 200,
-          "scaledWidth": 200,
-          "selectors": Object {
-            "itemsCountAttribute": "itemsSelector",
-            "renderComplete": "renderedSelector",
-            "renderError": "[dataRenderErrorSelector]",
-            "renderErrorAttribute": "dataRenderErrorSelector",
-            "screenshot": "screenshotSelector",
-            "timefilterDurationAttribute": "timefilterDurationSelector",
-          },
-          "useReportingBranding": true,
-          "width": 100,
-        },
-        "results": Array [
-          Object {
-            "elementsPositionAndAttributes": Array [
-              Object {
-                "attributes": Object {
-                  "description": "Default ",
-                  "title": "Default Mock Title",
-                },
-                "position": Object {
-                  "boundingClientRect": Object {
-                    "height": 600,
-                    "left": 0,
-                    "top": 0,
-                    "width": 800,
-                  },
-                  "scroll": Object {
-                    "x": 0,
-                    "y": 0,
-                  },
-                },
-              },
-            ],
-            "error": undefined,
-            "screenshots": Array [
-              Object {
-                "data": Object {
-                  "data": Array [
-                    115,
-                    99,
-                    114,
-                    101,
-                    101,
-                    110,
-                    115,
-                    104,
-                    111,
-                    116,
-                  ],
-                  "type": "Buffer",
-                },
-                "description": "Default ",
-                "title": "Default Mock Title",
-              },
-            ],
-            "timeRange": "Default GetTimeRange Result",
-          },
-        ],
-      }
-    `);
-  });
-
-  it('pipelines multiple urls into', async () => {
-    driver.screenshot.mockResolvedValue(Buffer.from('some screenshots'));
-
-    await expect(
-      getScreenshots(driverFactory, logger, {
-        ...options,
-        urls: ['/welcome/home/start/index2.htm', '/welcome/home/start/index.php3?page=./home.php'],
-      }).toPromise()
-    ).resolves.toMatchInlineSnapshot(`
-            Object {
-              "layout": PreserveLayout {
-                "groupCount": 1,
-                "hasFooter": true,
-                "hasHeader": true,
-                "height": 100,
-                "id": "preserve_layout",
-                "scaledHeight": 200,
-                "scaledWidth": 200,
-                "selectors": Object {
-                  "itemsCountAttribute": "itemsSelector",
-                  "renderComplete": "renderedSelector",
-                  "renderError": "[dataRenderErrorSelector]",
-                  "renderErrorAttribute": "dataRenderErrorSelector",
-                  "screenshot": "screenshotSelector",
-                  "timefilterDurationAttribute": "timefilterDurationSelector",
-                },
-                "useReportingBranding": true,
-                "width": 100,
-              },
-              "results": Array [
-                Object {
-                  "elementsPositionAndAttributes": Array [
-                    Object {
-                      "attributes": Object {
-                        "description": "Default ",
-                        "title": "Default Mock Title",
-                      },
-                      "position": Object {
-                        "boundingClientRect": Object {
-                          "height": 600,
-                          "left": 0,
-                          "top": 0,
-                          "width": 800,
-                        },
-                        "scroll": Object {
-                          "x": 0,
-                          "y": 0,
-                        },
-                      },
-                    },
-                  ],
-                  "error": undefined,
-                  "screenshots": Array [
-                    Object {
-                      "data": Object {
-                        "data": Array [
-                          115,
-                          111,
-                          109,
-                          101,
-                          32,
-                          115,
-                          99,
-                          114,
-                          101,
-                          101,
-                          110,
-                          115,
-                          104,
-                          111,
-                          116,
-                          115,
-                        ],
-                        "type": "Buffer",
-                      },
-                      "description": "Default ",
-                      "title": "Default Mock Title",
-                    },
-                  ],
-                  "timeRange": "Default GetTimeRange Result",
-                },
-                Object {
-                  "elementsPositionAndAttributes": Array [
-                    Object {
-                      "attributes": Object {
-                        "description": "Default ",
-                        "title": "Default Mock Title",
-                      },
-                      "position": Object {
-                        "boundingClientRect": Object {
-                          "height": 600,
-                          "left": 0,
-                          "top": 0,
-                          "width": 800,
-                        },
-                        "scroll": Object {
-                          "x": 0,
-                          "y": 0,
-                        },
-                      },
-                    },
-                  ],
-                  "error": undefined,
-                  "screenshots": Array [
-                    Object {
-                      "data": Object {
-                        "data": Array [
-                          115,
-                          111,
-                          109,
-                          101,
-                          32,
-                          115,
-                          99,
-                          114,
-                          101,
-                          101,
-                          110,
-                          115,
-                          104,
-                          111,
-                          116,
-                          115,
-                        ],
-                        "type": "Buffer",
-                      },
-                      "description": "Default ",
-                      "title": "Default Mock Title",
-                    },
-                  ],
-                  "timeRange": "Default GetTimeRange Result",
-                },
-              ],
-            }
-          `);
-
-    expect(driver.open).toHaveBeenCalledTimes(2);
-    expect(driver.open).nthCalledWith(
-      1,
-      expect.anything(),
-      expect.objectContaining({ waitForSelector: '.kbnAppWrapper' }),
-      expect.anything()
-    );
-    expect(driver.open).nthCalledWith(
-      2,
-      expect.anything(),
-      expect.objectContaining({ waitForSelector: '[data-shared-page="2"]' }),
-      expect.anything()
-    );
-  });
-
-  describe('error handling', () => {
-    it('recovers if waitForSelector fails', async () => {
-      driver.waitForSelector.mockImplementation((selectorArg: string) => {
-        throw new Error('Mock error!');
-      });
-      await expect(
-        getScreenshots(driverFactory, logger, {
-          ...options,
-          urls: [
-            '/welcome/home/start/index2.htm',
-            '/welcome/home/start/index.php3?page=./home.php3',
-          ],
-        }).toPromise()
-      ).resolves.toMatchInlineSnapshot(`
-              Object {
-                "layout": PreserveLayout {
-                  "groupCount": 1,
-                  "hasFooter": true,
-                  "hasHeader": true,
-                  "height": 100,
-                  "id": "preserve_layout",
-                  "scaledHeight": 200,
-                  "scaledWidth": 200,
-                  "selectors": Object {
-                    "itemsCountAttribute": "itemsSelector",
-                    "renderComplete": "renderedSelector",
-                    "renderError": "[dataRenderErrorSelector]",
-                    "renderErrorAttribute": "dataRenderErrorSelector",
-                    "screenshot": "screenshotSelector",
-                    "timefilterDurationAttribute": "timefilterDurationSelector",
-                  },
-                  "useReportingBranding": true,
-                  "width": 100,
-                },
-                "results": Array [
-                  Object {
-                    "elementsPositionAndAttributes": Array [
-                      Object {
-                        "attributes": Object {},
-                        "position": Object {
-                          "boundingClientRect": Object {
-                            "height": 100,
-                            "left": 0,
-                            "top": 0,
-                            "width": 100,
-                          },
-                          "scroll": Object {
-                            "x": 0,
-                            "y": 0,
-                          },
-                        },
-                      },
-                    ],
-                    "error": [Error: The "wait for elements" phase encountered an error: Error: An error occurred when trying to read the page for visualization panel info: Error: Mock error!],
-                    "screenshots": Array [
-                      Object {
-                        "data": Object {
-                          "data": Array [
-                            115,
-                            99,
-                            114,
-                            101,
-                            101,
-                            110,
-                            115,
-                            104,
-                            111,
-                            116,
-                          ],
-                          "type": "Buffer",
-                        },
-                        "description": undefined,
-                        "title": undefined,
-                      },
-                    ],
-                    "timeRange": null,
-                  },
-                  Object {
-                    "elementsPositionAndAttributes": Array [
-                      Object {
-                        "attributes": Object {},
-                        "position": Object {
-                          "boundingClientRect": Object {
-                            "height": 100,
-                            "left": 0,
-                            "top": 0,
-                            "width": 100,
-                          },
-                          "scroll": Object {
-                            "x": 0,
-                            "y": 0,
-                          },
-                        },
-                      },
-                    ],
-                    "error": [Error: The "wait for elements" phase encountered an error: Error: An error occurred when trying to read the page for visualization panel info: Error: Mock error!],
-                    "screenshots": Array [
-                      Object {
-                        "data": Object {
-                          "data": Array [
-                            115,
-                            99,
-                            114,
-                            101,
-                            101,
-                            110,
-                            115,
-                            104,
-                            111,
-                            116,
-                          ],
-                          "type": "Buffer",
-                        },
-                        "description": undefined,
-                        "title": undefined,
-                      },
-                    ],
-                    "timeRange": null,
-                  },
-                ],
-              }
-            `);
+      const testPipeline = () => test$.toPromise();
+      await expect(testPipeline).rejects.toMatchInlineSnapshot(
+        `[Error: The "Test Config" phase took longer than 0.2 seconds.]`
+      );
     });
 
-    it('observes page exit', async () => {
-      driverFactory.createPage.mockReturnValue(
-        of({ driver, exit$: throwError('Instant timeout has fired!') })
+    it('catches other Errors and explains where they were thrown', async () => {
+      const screenshots = new ScreenshotObservableHandler(browser, logger, layout, options);
+      const test$ = throwError(new Error(`Test Error to Throw`)).pipe(
+        screenshots.waitUntil(200, 'Test Config')
       );
 
-      await expect(
-        getScreenshots(driverFactory, logger, options).toPromise()
-      ).rejects.toMatchInlineSnapshot(`"Instant timeout has fired!"`);
+      const testPipeline = () => test$.toPromise();
+      await expect(testPipeline).rejects.toMatchInlineSnapshot(
+        `[Error: The "Test Config" phase encountered an error: Error: Test Error to Throw]`
+      );
     });
 
-    it(`uses defaults for element positions and size when Kibana page is not ready`, async () => {
-      driver.evaluate.mockImplementation(async (_, { context }) =>
-        context === CONTEXT_ELEMENTATTRIBUTES ? null : undefined
+    it('is a pass-through if there is no Error', async () => {
+      const screenshots = new ScreenshotObservableHandler(browser, logger, layout, options);
+      const test$ = of('nice to see you').pipe(screenshots.waitUntil(20, 'xxxxxxxxxxx'));
+
+      await expect(test$.toPromise()).resolves.toBe(`nice to see you`);
+    });
+  });
+
+  describe('checkPageIsOpen', () => {
+    it('throws a decorated Error when page is not open', async () => {
+      browser.isPageOpen.mockReturnValue(false);
+      const screenshots = new ScreenshotObservableHandler(browser, logger, layout, options);
+      const test$ = of(234455).pipe(
+        map((input) => {
+          screenshots.checkPageIsOpen();
+          return input;
+        })
       );
 
-      layout.getViewport = () => null;
+      await expect(test$.toPromise()).rejects.toMatchInlineSnapshot(
+        `[Error: Browser was closed unexpectedly! Check the server logs for more info.]`
+      );
+    });
 
-      await expect(getScreenshots(driverFactory, logger, options).toPromise()).resolves
-        .toMatchInlineSnapshot(`
-              Object {
-                "layout": PreserveLayout {
-                  "getViewport": [Function],
-                  "groupCount": 1,
-                  "hasFooter": true,
-                  "hasHeader": true,
-                  "height": 100,
-                  "id": "preserve_layout",
-                  "scaledHeight": 200,
-                  "scaledWidth": 200,
-                  "selectors": Object {
-                    "itemsCountAttribute": "itemsSelector",
-                    "renderComplete": "renderedSelector",
-                    "renderError": "[dataRenderErrorSelector]",
-                    "renderErrorAttribute": "dataRenderErrorSelector",
-                    "screenshot": "screenshotSelector",
-                    "timefilterDurationAttribute": "timefilterDurationSelector",
-                  },
-                  "useReportingBranding": true,
-                  "width": 100,
-                },
-                "results": Array [
-                  Object {
-                    "elementsPositionAndAttributes": Array [
-                      Object {
-                        "attributes": Object {},
-                        "position": Object {
-                          "boundingClientRect": Object {
-                            "height": 1200,
-                            "left": 0,
-                            "top": 0,
-                            "width": 1800,
-                          },
-                          "scroll": Object {
-                            "x": 0,
-                            "y": 0,
-                          },
-                        },
-                      },
-                    ],
-                    "error": undefined,
-                    "screenshots": Array [
-                      Object {
-                        "data": Object {
-                          "data": Array [
-                            115,
-                            99,
-                            114,
-                            101,
-                            101,
-                            110,
-                            115,
-                            104,
-                            111,
-                            116,
-                          ],
-                          "type": "Buffer",
-                        },
-                        "description": undefined,
-                        "title": undefined,
-                      },
-                    ],
-                    "timeRange": undefined,
-                  },
-                ],
-              }
-            `);
+    it('is a pass-through when the page is open', async () => {
+      const screenshots = new ScreenshotObservableHandler(browser, logger, layout, options);
+      const test$ = of(234455).pipe(
+        map((input) => {
+          screenshots.checkPageIsOpen();
+          return input;
+        })
+      );
+
+      await expect(test$.toPromise()).resolves.toBe(234455);
     });
   });
 });
