--- x-pack/plugins/screenshotting/server/screenshots/observable_handler.ts
@@ -1,211 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import apm from 'elastic-apm-node';
-import { defer, forkJoin, throwError, Observable } from 'rxjs';
-import { catchError, mergeMap, switchMapTo, timeoutWith } from 'rxjs/operators';
-import type { Logger } from 'src/core/server';
-import type { Layout as ScreenshotModeLayout } from 'src/plugins/screenshot_mode/common';
-import type { ConditionalHeaders, HeadlessChromiumDriver } from '../browsers';
-import { getChromiumDisconnectedError } from '../browsers';
-import type { Layout } from '../layouts';
-import type { ElementsPositionAndAttribute } from './get_element_position_data';
-import { getElementPositionAndAttributes } from './get_element_position_data';
-import { getNumberOfItems } from './get_number_of_items';
-import { getRenderErrors } from './get_render_errors';
-import { getScreenshots } from './get_screenshots';
-import type { Screenshot } from './get_screenshots';
-import { getTimeRange } from './get_time_range';
-import { injectCustomCss } from './inject_css';
-import { openUrl } from './open_url';
-import type { UrlOrUrlWithContext } from './open_url';
-import { waitForRenderComplete } from './wait_for_render';
-import { waitForVisualizations } from './wait_for_visualizations';
-
-export interface PhaseTimeouts {
-  openUrl: number;
-  waitForElements: number;
-  renderComplete: number;
-  loadDelay: number;
-}
-
-export interface ScreenshotObservableOptions {
-  browserTimezone?: string;
-  conditionalHeaders: ConditionalHeaders;
-  timeouts: PhaseTimeouts;
-  urls: UrlOrUrlWithContext[];
-}
-
-export interface ScreenshotObservableResult {
-  timeRange: string | null;
-  screenshots: Screenshot[];
-  error?: Error;
-
-  /**
-   * Individual visualizations might encounter errors at runtime. If there are any they are added to this
-   * field. Any text captured here is intended to be shown to the user for debugging purposes, reporting
-   * does no further sanitization on these strings.
-   */
-  renderErrors?: string[];
-  elementsPositionAndAttributes?: ElementsPositionAndAttribute[]; // NOTE: for testing
-}
-
-interface PageSetupResults {
-  elementsPositionAndAttributes: ElementsPositionAndAttribute[] | null;
-  timeRange: string | null;
-  error?: Error;
-}
-
-const DEFAULT_SCREENSHOT_CLIP_HEIGHT = 1200;
-const DEFAULT_SCREENSHOT_CLIP_WIDTH = 1800;
-
-const getDefaultElementPosition = (dimensions: { height?: number; width?: number } | null) => {
-  const height = dimensions?.height || DEFAULT_SCREENSHOT_CLIP_HEIGHT;
-  const width = dimensions?.width || DEFAULT_SCREENSHOT_CLIP_WIDTH;
-
-  return [
-    {
-      position: {
-        boundingClientRect: { top: 0, left: 0, height, width },
-        scroll: { x: 0, y: 0 },
-      },
-      attributes: {},
-    },
-  ];
-};
-
-/*
- * If Kibana is showing a non-HTML error message, the viewport might not be
- * provided by the browser.
- */
-const getDefaultViewPort = () => ({
-  height: DEFAULT_SCREENSHOT_CLIP_HEIGHT,
-  width: DEFAULT_SCREENSHOT_CLIP_WIDTH,
-  zoom: 1,
-});
-
-export class ScreenshotObservableHandler {
-  constructor(
-    private readonly driver: HeadlessChromiumDriver,
-    private readonly logger: Logger,
-    private readonly layout: Layout,
-    private options: ScreenshotObservableOptions
-  ) {}
-
-  /*
-   * Decorates a TimeoutError with context of the phase that has timed out.
-   */
-  public waitUntil<O>(timeoutValue: number, label: string) {
-    return (source: Observable<O>) =>
-      source.pipe(
-        catchError((error) => {
-          throw new Error(`The "${label}" phase encountered an error: ${error}`);
-        }),
-        timeoutWith(
-          timeoutValue,
-          throwError(
-            new Error(`The "${label}" phase took longer than ${timeoutValue / 1000} seconds.`)
-          )
-        )
-      );
-  }
-
-  private openUrl(index: number, url: UrlOrUrlWithContext) {
-    return defer(() =>
-      openUrl(
-        this.driver,
-        this.logger,
-        this.options.timeouts.openUrl,
-        index,
-        url,
-        this.options.conditionalHeaders,
-        this.layout.id as ScreenshotModeLayout
-      )
-    ).pipe(this.waitUntil(this.options.timeouts.openUrl, 'open URL'));
-  }
-
-  private waitForElements() {
-    const driver = this.driver;
-    const waitTimeout = this.options.timeouts.waitForElements;
-
-    return defer(() => getNumberOfItems(driver, this.logger, waitTimeout, this.layout)).pipe(
-      mergeMap((itemsCount) => {
-        // set the viewport to the dimentions from the job, to allow elements to flow into the expected layout
-        const viewport = this.layout.getViewport(itemsCount) || getDefaultViewPort();
-
-        return forkJoin([
-          driver.setViewport(viewport, this.logger),
-          waitForVisualizations(driver, this.logger, waitTimeout, itemsCount, this.layout),
-        ]);
-      }),
-      this.waitUntil(waitTimeout, 'wait for elements')
-    );
-  }
-
-  private completeRender(apmTrans: apm.Transaction | null) {
-    const driver = this.driver;
-    const layout = this.layout;
-    const logger = this.logger;
-
-    return defer(async () => {
-      // Waiting till _after_ elements have rendered before injecting our CSS
-      // allows for them to be displayed properly in many cases
-      await injectCustomCss(driver, logger, layout);
-
-      const apmPositionElements = apmTrans?.startSpan('position-elements', 'correction');
-      // position panel elements for print layout
-      await layout.positionElements?.(driver, logger);
-      apmPositionElements?.end();
-
-      await waitForRenderComplete(driver, logger, this.options.timeouts.loadDelay, layout);
-    }).pipe(
-      mergeMap(() =>
-        forkJoin({
-          timeRange: getTimeRange(driver, logger, layout),
-          elementsPositionAndAttributes: getElementPositionAndAttributes(driver, logger, layout),
-          renderErrors: getRenderErrors(driver, logger, layout),
-        })
-      ),
-      this.waitUntil(this.options.timeouts.renderComplete, 'render complete')
-    );
-  }
-
-  public setupPage(index: number, url: UrlOrUrlWithContext, apmTrans: apm.Transaction | null) {
-    return this.openUrl(index, url).pipe(
-      switchMapTo(this.waitForElements()),
-      switchMapTo(this.completeRender(apmTrans))
-    );
-  }
-
-  public getScreenshots() {
-    return (withRenderComplete: Observable<PageSetupResults>) =>
-      withRenderComplete.pipe(
-        mergeMap(async (data: PageSetupResults): Promise<ScreenshotObservableResult> => {
-          this.checkPageIsOpen(); // fail the report job if the browser has closed
-
-          const elements =
-            data.elementsPositionAndAttributes ??
-            getDefaultElementPosition(this.layout.getViewport(1));
-          const screenshots = await getScreenshots(this.driver, this.logger, elements);
-          const { timeRange, error: setupError } = data;
-
-          return {
-            timeRange,
-            screenshots,
-            error: setupError,
-            elementsPositionAndAttributes: elements,
-          };
-        })
-      );
-  }
-
-  public checkPageIsOpen() {
-    if (!this.driver.isPageOpen()) {
-      throw getChromiumDisconnectedError();
-    }
-  }
-}
